+++ src/rseqc/rseqc_junctionsaturation/script.sh
@@ -15,17 +15,5 @@ junction_saturation.py \
     -v $par_min_splice_read \
     -q $par_map_qual
 
-mv $prefix.junctionSaturation_plot.pdf $par_output_plot
-mv $prefix.junctionSaturation_plot.r $par_output_plot_r
-
-# Version
-
-text="${meta_functionality_name}:
-    rseqc: $(junction_saturation.py --version | sed -e 's/junction_saturation.py //g')"
-
-if [ -e "$par_versions" ]; then
-    echo "$text" >> "$par_versions"
-    mv "$par_versions" "$par_updated_versions"
-else
-    echo "$text" > "$par_updated_versions"
-fi
+[[ -f "$prefix.junctionSaturation_plot.pdf" ]] && mv $prefix.junctionSaturation_plot.pdf $par_output_plot
+[[ -f "$prefix.junctionSaturation_plot.r" ]] && mv $prefix.junctionSaturation_plot.r $par_output_plot_r
