+++ src/rseqc/tin/script.sh
@@ -23,18 +23,7 @@ tin.py \
     -n $par_sample_size \
     -s $par_subtract_background
 
-mv ${bam_file%.*}.summary.txt $par_output_tin_summary
-mv ${bam_file%.*}.tin.xls $par_output_tin
+[[ -f "${bam_file%.*}.summary.txt" ]] && mv ${bam_file%.*}.summary.txt $par_output_tin_summary
+[[ -f "${bam_file%.*}.tin.xls" ]] && mv ${bam_file%.*}.tin.xls $par_output_tin
 
 clean_up
-
-# Version
-text="${meta_functionality_name}:
-    rseqc: $(tin.py --version | sed -e 's/tin.py //g')"
-
-if [ -e "$par_versions" ]; then
-    echo "$text" >> "$par_versions"
-    mv "$par_versions" "$par_updated_versions"
-else
-    echo "$text" > "$par_updated_versions"
-fi
